Moved `Direction8` into game's code
