+++ src/GameLogic/Enums/Direction8.ts
