+++ test/GameLogic/e2e/floorInteraction/WaterFloor.test.ts
@@ -4,7 +4,7 @@ import {FloorType, PlayerAction, EntityType} from '../../../../src/GameLogic/Enu
 import {SessionPlayer} from '../../helpers/SessionPlayer';
 import {TestLevelBuilder} from '../../helpers/TestLevelBuilder';
 import {Fireball} from '../../../../src/GameLogic/Entities/Fireball';
-import {Direction8} from '../../../../src.common/Enums/Direction8';
+import {Direction8} from '../../../../src/GameLogic/Enums/Direction8';
 
 describe('GameLogic.e2e - Water floor', () => {
 	it('Protagonist should be removed when moving onto water', () => {
